backend/pyproject.toml
[project]
name = "ts-teamtakt-backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29",
    "alembic>=1.13",
    "pydantic-settings>=2.5",
    "python-dotenv>=1.0",
    "python-jose[cryptography]>=3.3",
    "httpx>=0.27",
    "openpyxl>=3.1",
    "pytz>=2024.1",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-mock>=3.12",
    "httpx>=0.27",
    "aiosqlite>=0.20",
    "ruff>=0.15.9",
    "pyright>=1.1.408",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["app/tests"]

[tool.ruff]
target-version = "py312"
line-length = 88
exclude = ["alembic/versions/"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["E501"]

[tool.pyright]
pythonVersion = "3.12"
venvPath = "."
venv = ".venv"
typeCheckingMode = "basic"
include = ["app"]
exclude = ["alembic"]
